Portfolio improvements (#3924)
* Closes FE-1188

* Closes FE-1009

* lowercase open orders

* hide nav bar title

* optimize rendering

* SSR swap tool

* clean up limit order ff & bg

* remove old swap tool code

* add knip JSON

* fix translations test

* coderabbit review
