+++ packages/web/components/complex/portfolio/cypher-card.tsx
@@ -1,4 +1,5 @@
 import Link from "next/link";
+import { useLocalStorage } from "react-use";
 
 import { Icon } from "~/components/assets";
 import { useTranslation } from "~/hooks";
@@ -7,9 +8,18 @@ export const CYPHER_CARD_URL = "https://pay.osmosis.zone?ref=osmosis";
 
 export const CypherCard = () => {
   const { t } = useTranslation();
+  const [isClosed, setIsClosed] = useLocalStorage("cypher-card-closed", false);
+
+  if (isClosed) return null;
+
   return (
-    <Link href={CYPHER_CARD_URL} target="_blank" rel="noopener noreferrer">
-      <div className="flex h-[4.5rem] w-full flex-1 items-center gap-4 rounded-2xl bg-osmoverse-850 px-4 py-3">
+    <Link
+      href={CYPHER_CARD_URL}
+      target="_blank"
+      rel="noopener noreferrer"
+      referrerPolicy="no-referrer"
+    >
+      <div className="relative flex h-[4.5rem] w-full flex-1 items-center gap-2 rounded-2xl bg-osmoverse-850 px-4 py-3">
         <div className="mr-3 flex h-12 w-12 flex-col items-center justify-center rounded-xl bg-osmoverse-800">
           <Icon id="cypher-card" />
         </div>
@@ -21,9 +31,20 @@ export const CypherCard = () => {
             {t("cypherCard.cypherOrder")}
           </p>
         </div>
-        <div className="caption ml-auto rounded-xl bg-osmoverse-800 p-1 px-2 text-osmoverse-300">
+
+        <div className="caption rounded-xl bg-osmoverse-800 p-1 px-2 text-osmoverse-300">
           {t("cypherCard.cypherBeta")}
         </div>
+
+        <button
+          onClick={(e) => {
+            e.preventDefault();
+            setIsClosed(true);
+          }}
+          className="absolute right-2 top-1/2 -translate-y-1/2 p-1 text-osmoverse-300 hover:text-white-full"
+        >
+          <Icon id="close" width={16} height={16} />
+        </button>
       </div>
     </Link>
   );
