+++ packages/web/components/complex/portfolio/portfolio-page.tsx
@@ -2,7 +2,6 @@ import { Tab, TabGroup, TabList, TabPanel, TabPanels } from "@headlessui/react";
 import classNames from "classnames";
 import { observer } from "mobx-react-lite";
 import { FunctionComponent } from "react";
-import { useLocalStorage } from "react-use";
 
 import { Allocation } from "~/components/complex/portfolio/allocation";
 import { AssetsOverview } from "~/components/complex/portfolio/assets-overview";
@@ -22,19 +21,20 @@ import {
   useWalletSelect,
 } from "~/hooks";
 import { useStore } from "~/stores";
+import { HideDustState } from "~/stores/user-settings/hide-dust";
 import { api } from "~/utils/trpc";
 
 import { CypherCard } from "./cypher-card";
 import { GetStartedWithOsmosis } from "./get-started-with-osmosis";
-import { PORTFOLIO_HIDE_DUST_KEY } from "./portfolio-dust";
 
 export const PortfolioPage: FunctionComponent = observer(() => {
   const { t } = useTranslation();
-  const { accountStore } = useStore();
+  const { accountStore, userSettings } = useStore();
   const wallet = accountStore.getWallet(accountStore.osmosisChainId);
+  const { isLoading: isWalletLoading } = useWalletSelect();
   const featureFlags = useFeatureFlags();
 
-  useAmplitudeAnalytics({
+  const { logEvent } = useAmplitudeAnalytics({
     onLoadEvent: [EventName.Portfolio.pageViewed],
   });
 
@@ -53,25 +53,18 @@ export const PortfolioPage: FunctionComponent = observer(() => {
 
   const totalCap = allocation?.totalCap;
 
-  const userHasNoAssets = allocation && totalCap?.toDec()?.isZero();
+  const userHasNoAssets = Boolean(allocation && totalCap?.toDec()?.isZero());
 
   const [overviewRef, { height: overviewHeight }] =
     useDimension<HTMLDivElement>();
   const [tabsRef, { height: tabsHeight }] = useDimension<HTMLDivElement>();
 
-  const { logEvent } = useAmplitudeAnalytics();
-
   const isWalletConnected =
     wallet && wallet.isWalletConnected && wallet.address;
 
-  const { isLoading: isWalletLoading } = useWalletSelect();
-
-  const [hideDust, setHideDust] = useLocalStorage(
-    PORTFOLIO_HIDE_DUST_KEY,
-    true
-  );
-  const showZeroBalancesSplash =
-    userHasNoAssets === true || userHasNoAssets === undefined;
+  const hideDustSettingStore =
+    userSettings.getUserSettingById<HideDustState>("hide-dust");
+  const hideDust = Boolean(hideDustSettingStore?.state?.hideDust);
 
   return (
     <div className="flex justify-center p-8 pt-4 1.5xl:flex-col md:p-4">
@@ -130,14 +123,16 @@ export const PortfolioPage: FunctionComponent = observer(() => {
                   <div className="mx-auto my-6 w-fit">
                     <Spinner />
                   </div>
-                ) : showZeroBalancesSplash ? (
+                ) : userHasNoAssets ? (
                   <UserZeroBalanceTableSplash />
                 ) : (
                   <TabPanels>
                     <TabPanel>
                       <PortfolioAssetBalancesTable
                         hideDust={Boolean(hideDust)}
-                        setHideDust={setHideDust}
+                        setHideDust={(hideDust) =>
+                          hideDustSettingStore?.setState({ hideDust })
+                        }
                         tableTopPadding={overviewHeight + tabsHeight}
                       />
                     </TabPanel>
@@ -166,10 +161,7 @@ export const PortfolioPage: FunctionComponent = observer(() => {
             >
               {featureFlags.cypherCard && <CypherCard />}
               {!isLoadingAllocation && !userHasNoAssets && (
-                <Allocation
-                  allocation={allocation}
-                  hideDust={Boolean(hideDust)}
-                />
+                <Allocation allocation={allocation} />
               )}
             </div>
             <div className="flex w-full flex-col">
