+++ packages/web/hooks/use-feature-flags.ts
@@ -1,11 +1,8 @@
 import { AvailableFlags } from "@osmosis-labs/types";
-import { apiClient } from "@osmosis-labs/utils";
-import { useQuery } from "@tanstack/react-query";
 import { useFlags, useLDClient } from "launchdarkly-react-client-sdk";
 import { useEffect, useState } from "react";
 
 import { useWindowSize } from "~/hooks";
-import { LevanaGeoBlockedResponse } from "~/pages/_app";
 
 const defaultFlags: Record<AvailableFlags, boolean> = {
   staking: true,
@@ -37,27 +34,11 @@ const defaultFlags: Record<AvailableFlags, boolean> = {
   nomicWithdrawAmount: false,
 };
 
-const LIMIT_ORDER_COUNTRY_CODES =
-  process.env.NEXT_PUBLIC_LIMIT_ORDER_COUNTRY_CODES?.split(",").map((s) =>
-    s.trim()
-  ) ?? [];
-
 export function useFeatureFlags() {
   const launchdarklyFlags: Record<AvailableFlags, boolean> = useFlags();
   const { isMobile } = useWindowSize();
   const [isInitialized, setIsInitialized] = useState(false);
   const client = useLDClient();
-
-  const { data: levanaGeoblock } = useQuery(
-    ["levana-geoblocked"],
-    () =>
-      apiClient<LevanaGeoBlockedResponse>("https://geoblocked.levana.finance/"),
-    {
-      staleTime: Infinity,
-      cacheTime: Infinity,
-      retry: false,
-    }
-  );
 
   useEffect(() => {
     if (!isInitialized && client && process.env.NODE_ENV !== "test")
@@ -86,18 +67,6 @@ export function useFeatureFlags() {
         launchdarklyFlags.oneClickTrading,
     _isInitialized: isDevModeWithoutClientID ? true : isInitialized,
     _isClientIDPresent: !!process.env.NEXT_PUBLIC_LAUNCH_DARKLY_CLIENT_SIDE_ID,
-    limitOrders: isDevModeWithoutClientID
-      ? defaultFlags.limitOrders
-      : isInitialized &&
-        launchdarklyFlags.limitOrders &&
-        (LIMIT_ORDER_COUNTRY_CODES.length === 0 ||
-          LIMIT_ORDER_COUNTRY_CODES.includes(
-            levanaGeoblock?.countryCode ?? ""
-          )),
-    // To test chain upgrades easily on Edgenet, uncomment the flags below
-    // limitOrders: true,
-    // oneClickTrading: true,
-    // staking: true,
   } as Record<
     AvailableFlags | "_isInitialized" | "_isClientIDPresent",
     boolean
