+++ packages/web/knip.json
@@ -1,0 +1,3 @@
+{
+  "ignore": [".next/**/*", "**/tradingview/**"]
+}
