+++ packages/web/localizations/es.json
@@ -329,14 +329,6 @@
   },
   "estimated": "Estimado",
   "frontierMigration": {
-    "introducingUnverifiedAssets": "Introduciendo Activos No Verificados",
-    "simplifiedExperience": "¡Tu experiencia ha sido simplificada!",
-    "frontierHasNowMerged": "La página de Frontier ahora se ha fusionado con",
-    "thisMeansManaging": "¡Esto significa que la gestión de tokens se puede hacer cómodamente en un solo lugar!",
-    "commitmentToDecentralization": "En línea con el compromiso con la descentralización y la accesibilidad, la opción de manejar tokens no verificados aún está disponible.",
-    "settingIsNowEnabled": "La configuración ahora está habilitada.",
-    "youMayDisable": "Puedes desactivar esta configuración en cualquier momento en el modal de configuraciones.",
-    "openSettings": "Abrir configuraciones",
     "proceed": "Proceder"
   },
   "lockToken": {
@@ -704,19 +696,10 @@
       "hide": "Ocultar",
       "show": "Mostrar"
     },
-    "maxButtonErrorNoBalance": "Saldo insuficiente para calcular el importe máximo.",
-    "available": "Disponible",
     "button": "Intercambio",
     "buttonError": "Intercambio de todos modos",
-    "expectedOutput": "Rendimiento esperado",
-    "fee": "Tarifa de intercambio ({fee})",
-    "networkFee": "Tarifa de red",
-    "totalFee": "Costo total",
     "HALF": "Mitad",
-    "MAX": "Máximo",
-    "minimumSlippage": "Mínimo recibido después del deslizamiento ({slippage})",
     "pool": "Piscina #{id}",
-    "priceImpact": "Impacto en el mercado",
     "routerTooltipFee": "Tarifa",
     "routerTooltipSpreadFactor": "Factor de dispersión",
     "showDetails": "Mostrar detalles",
@@ -727,9 +710,7 @@
       "increaseSpendLimit": "Aumentar el límite de gasto"
     },
     "settings": {
-      "slippage": "Tolerancia de deslizamiento",
-      "slippageInfo": "Su transacción se revertirá si el precio cambia desfavorablemente en más de este porcentaje.",
-      "title": "Configuración de transacción"
+      "slippage": "Tolerancia de deslizamiento"
     },
     "title": "Intercambio",
     "dynamicSpreadFactor": "Dinámica",
