+++ packages/web/localizations/gu.json
@@ -329,14 +329,6 @@
   },
   "estimated": "અંદાજિત",
   "frontierMigration": {
-    "introducingUnverifiedAssets": "વણચકાસાયેલ અસ્કયામતોનો પરિચય",
-    "simplifiedExperience": "તમારો અનુભવ સરળ કરવામાં આવ્યો છે!",
-    "frontierHasNowMerged": "ફ્રન્ટિયર પેજ હવે સાથે મર્જ થઈ ગયું છે",
-    "thisMeansManaging": "આનો અર્થ એ છે કે ટોકન્સનું સંચાલન એક જગ્યાએ સરળતાથી કરી શકાય છે!",
-    "commitmentToDecentralization": "વિકેન્દ્રીકરણ અને સુલભતા માટેની પ્રતિબદ્ધતા અનુસાર, વણચકાસાયેલ ટોકન્સને હેન્ડલ કરવાનો વિકલ્પ હજુ પણ ઉપલબ્ધ છે.",
-    "settingIsNowEnabled": "સેટિંગ હવે સક્ષમ છે.",
-    "youMayDisable": "તમે સેટિંગ્સ મોડલમાં કોઈપણ સમયે આ સેટિંગને અક્ષમ કરી શકો છો.",
-    "openSettings": "સેટિંગ્સ ખોલો",
     "proceed": "આગળ વધો"
   },
   "lockToken": {
@@ -704,19 +696,10 @@
       "hide": "છુપાવો",
       "show": "બતાવો"
     },
-    "maxButtonErrorNoBalance": "મહત્તમ રકમની ગણતરી કરવા માટે પર્યાપ્ત બેલેન્સ નથી.",
-    "available": "ઉપલબ્ધ છે",
     "button": "સ્વેપ",
     "buttonError": "કોઈપણ રીતે સ્વેપ કરો",
-    "expectedOutput": "અપેક્ષિત આઉટપુટ",
-    "fee": "સ્વેપ ફી ( {fee} )",
-    "networkFee": "નેટવર્ક ફી",
-    "totalFee": "કુલ ફી",
     "HALF": "અર્ધ",
-    "MAX": "MAX",
-    "minimumSlippage": "સ્લિપેજ પછી પ્રાપ્ત ન્યૂનતમ ( {slippage} )",
     "pool": "પૂલ # {id}",
-    "priceImpact": "બજારની અસર",
     "routerTooltipFee": "ફી",
     "routerTooltipSpreadFactor": "સ્પ્રેડ ફેક્ટર",
     "showDetails": "વિગતો બતાવો",
@@ -727,9 +710,7 @@
       "increaseSpendLimit": "ખર્ચ મર્યાદા વધારો"
     },
     "settings": {
-      "slippage": "સ્લિપેજ સહનશીલતા",
-      "slippageInfo": "જો કિંમત આ ટકાવારી કરતાં વધુ પ્રતિકૂળ રીતે બદલાશે તો તમારો વ્યવહાર પાછો આવશે.",
-      "title": "વ્યવહાર સેટિંગ્સ"
+      "slippage": "સ્લિપેજ સહનશીલતા"
     },
     "title": "સ્વેપ",
     "dynamicSpreadFactor": "ગતિશીલ",
