+++ packages/web/localizations/ja.json
@@ -329,14 +329,6 @@
   },
   "estimated": "推定",
   "frontierMigration": {
-    "introducingUnverifiedAssets": "未検証のアセットの紹介",
-    "simplifiedExperience": "あなたの経験は簡素化されました!",
-    "frontierHasNowMerged": "フロンティアページは次と統合されました",
-    "thisMeansManaging": "つまり、トークンの管理を 1 か所で簡単に行うことができます。",
-    "commitmentToDecentralization": "分散化とアクセシビリティへの取り組みに沿って、未検証のトークンを処理するオプションも引き続き利用できます。",
-    "settingIsNowEnabled": "これで設定が有効になりました。",
-    "youMayDisable": "この設定は、設定モーダルでいつでも無効にすることができます。",
-    "openSettings": "設定を開く",
     "proceed": "進む"
   },
   "lockToken": {
@@ -704,19 +696,10 @@
       "hide": "隠れる",
       "show": "見せる"
     },
-    "maxButtonErrorNoBalance": "最大額を計算するには残高が足りません。",
-    "available": "利用可能",
     "button": "スワップ",
     "buttonError": "とにかく交換する",
-    "expectedOutput": "期待される出力",
-    "fee": "スワップ手数料 ( {fee} )",
-    "networkFee": "ネットワーク料金",
-    "totalFee": "合計料金",
     "HALF": "半分",
-    "MAX": "マックス",
-    "minimumSlippage": "スリッページ後に受信した最小値 ( {slippage} )",
     "pool": "プール番号{id}",
-    "priceImpact": "市場への影響",
     "routerTooltipFee": "手数料",
     "routerTooltipSpreadFactor": "スプレッドファクター",
     "showDetails": "詳細を表示",
@@ -727,9 +710,7 @@
       "increaseSpendLimit": "支出制限を増やす"
     },
     "settings": {
-      "slippage": "滑り許容差",
-      "slippageInfo": "価格がこのパーセンテージを超えて不利に変化した場合、トランザクションは元に戻ります。",
-      "title": "トランザクション設定"
+      "slippage": "滑り許容差"
     },
     "title": "スワップ",
     "dynamicSpreadFactor": "動的",
