+++ packages/web/localizations/ru.json
@@ -329,14 +329,6 @@
   },
   "estimated": "Оцененный",
   "frontierMigration": {
-    "introducingUnverifiedAssets": "Представляем непроверенные активы",
-    "simplifiedExperience": "Ваш опыт был упрощен!",
-    "frontierHasNowMerged": "Страница Frontier теперь объединена с",
-    "thisMeansManaging": "Это означает, что управление токенами можно удобно осуществлять в одном месте!",
-    "commitmentToDecentralization": "В соответствии с приверженностью к децентрализации и доступности, возможность обработки непроверенных токенов по-прежнему доступна.",
-    "settingIsNowEnabled": "Теперь настройка включена.",
-    "youMayDisable": "Вы можете отключить этот параметр в любое время в модальном окне настроек.",
-    "openSettings": "Открыть настройки",
     "proceed": "Продолжить"
   },
   "lockToken": {
@@ -704,19 +696,10 @@
       "hide": "Скрывать",
       "show": "Показывать"
     },
-    "maxButtonErrorNoBalance": "Недостаточно баланса для расчета максимальной суммы.",
-    "available": "Доступный",
     "button": "Менять",
     "buttonError": "Все равно поменять",
-    "expectedOutput": "Ожидаемый результат",
-    "fee": "Комиссия за своп ( {fee} )",
-    "networkFee": "Сетевая плата",
-    "totalFee": "Общая плата",
     "HALF": "ПОЛОВИНА",
-    "MAX": "МАКС",
-    "minimumSlippage": "Минимум, полученный после проскальзывания ( {slippage} )",
     "pool": "Пул № {id}",
-    "priceImpact": "Влияние на рынок",
     "routerTooltipFee": "Платеж",
     "routerTooltipSpreadFactor": "Фактор распространения",
     "showDetails": "Показать детали",
@@ -727,9 +710,7 @@
       "increaseSpendLimit": "Увеличить лимит расходов"
     },
     "settings": {
-      "slippage": "Допуск проскальзывания",
-      "slippageInfo": "Ваша транзакция будет отменена, если цена изменится в неблагоприятную сторону более чем на этот процент.",
-      "title": "Настройки транзакции"
+      "slippage": "Допуск проскальзывания"
     },
     "title": "Менять",
     "dynamicSpreadFactor": "Динамический",
