+++ packages/web/localizations/scripts/omitted-keys.mjs
@@ -1,0 +1,6 @@
+/** Add key paths here to skip them in localizations tests. */
+export const omittedKeyPaths = [
+  "assets.categories",
+  "limitOrders.historyTable.columns",
+  "transfer.nomic.estimatedTime",
+];
