+++ packages/web/localizations/tr.json
@@ -227,7 +227,7 @@
     "seeAll": "Tümünü gör",
     "showHidden": "Gizliyi göster ( {hiddenDustCount} )",
     "hideSmallBalances": "Küçük bakiyeleri gizle",
-    "openOrders": "Açık Siparişler",
+    "openOrders": "Açık siparişler",
     "last24h": "Son 24 saat",
     "last7d": "Son 7 gün",
     "last30d": "Son 30 gün",
@@ -329,14 +329,6 @@
   },
   "estimated": "Tahmini",
   "frontierMigration": {
-    "introducingUnverifiedAssets": "Doğrulanmamış Varlıkların Tanıtılması",
-    "simplifiedExperience": "Deneyiminiz basitleştirildi!",
-    "frontierHasNowMerged": "Frontier sayfası şimdi ile birleşti",
-    "thisMeansManaging": "Bu, belirteçleri tek bir yerde uygun bir şekilde yönetebileceğiniz anlamına gelir!",
-    "commitmentToDecentralization": "Merkeziyetsizleşme ve erişilebilirlik taahhüdü ile, doğrulanmamış belirteçleri işleme seçeneği hala mevcut.",
-    "settingIsNowEnabled": "Ayar şimdi etkinleştirildi.",
-    "youMayDisable": "Bu ayarı istediğiniz zaman ayarlar modülünde devre dışı bırakabilirsiniz.",
-    "openSettings": "Ayarları Aç",
     "proceed": "Devam et"
   },
   "lockToken": {
@@ -704,19 +696,10 @@
       "hide": "Saklamak",
       "show": "Göstermek"
     },
-    "maxButtonErrorNoBalance": "Maksimum tutarı hesaplamak için yeterli bakiye yok.",
-    "available": "Mevcut",
     "button": "Takas",
     "buttonError": "Yine de takası gerçekleştir",
-    "expectedOutput": "Elde Edilmesi Beklenen Miktar",
-    "fee": "Takas Ücreti ({fee})",
-    "networkFee": "Ağ Ücreti",
-    "totalFee": "Toplam ücret",
     "HALF": "YARISI",
-    "MAX": "MAKS.",
-    "minimumSlippage": "Slipaj dahil elde edilecek minimum ({slippage})",
     "pool": "Havuz #{id}",
-    "priceImpact": "Pazar Etkisi",
     "routerTooltipFee": "Ücret",
     "routerTooltipSpreadFactor": "Yayılma Faktörü",
     "showDetails": "Detayları göster",
@@ -727,9 +710,7 @@
       "increaseSpendLimit": "Harcama sınırını artırın"
     },
     "settings": {
-      "slippage": "Slipaj toleransı",
-      "slippageInfo": "Fiyatta belirttiğiniz yüzdenin üzerinde artış olursa işleminiz iptal olacaktır.",
-      "title": "İşlem Ayarları"
+      "slippage": "Slipaj toleransı"
     },
     "title": "Takas",
     "dynamicSpreadFactor": "Dinamik",
