+++ packages/web/pages/assets/[denom].tsx
@@ -30,7 +30,7 @@ import {
 import { AssetNavigation } from "~/components/pages/asset-info-page/navigation";
 import { AssetPools } from "~/components/pages/asset-info-page/pools";
 import { TwitterSection } from "~/components/pages/asset-info-page/twitter";
-import { SwapToolProps } from "~/components/swap-tool/alt";
+import { SwapToolProps } from "~/components/swap-tool";
 import { TradeTool } from "~/components/trade-tool";
 import { EventName } from "~/config";
 import { AssetLists } from "~/config/generated/asset-lists";
