+++ packages/web/styles/globals.css
@@ -33,7 +33,7 @@
     @apply text-white-full;
   }
   body {
-    @apply bg-osmoverse-900;
+    @apply bg-osmoverse-1000;
   }
 
   h1 {
@@ -107,7 +107,7 @@
   }
 
   thead > tr {
-    @apply sticky top-navbar z-40 border-b border-osmoverse-700 bg-osmoverse-1000 text-body2 font-semibold text-osmoverse-300 md:top-navbar-mobile;
+    @apply sticky top-navbar z-40 border-b border-osmoverse-700 text-body2 font-semibold text-osmoverse-300 md:top-navbar-mobile;
   }
 
   thead > tr > th:first-child {
